requirements.txt
pyperclip>=1.6.0
requests>=2.18.4
adal>=0.4.7
azure-mgmt-authorization==0.51.1
azure-graphrbac==0.61.0